+++ src/components/content/MainContent.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const MainContent = () => {
+    return (
+        <div>
+            <h1>Reasons I'm excited to learn React</h1>
+            <ol>
+                <li>Something refreshing</li>
+                <li>Hireable skill</li>
+                <li>Broad horizons</li>
+            </ol>
+        </div>
+    )
+}
+
+export default MainContent
