+++ src/components/layout/Footer.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Footer = () => {
+    return (
+        <footer className="footer">
+            <small>
+                &copy; 2023 MikiWojak development. All rights reserved.
+            </small>
+        </footer>
+    )
+}
+
+export default Footer
